@@ -20,6 +20,7 @@
     "@three-ts/orbit-controls": "1.4.7",
     "axios": "^1.7.3",
     "buffer": "^6.0.3",
+    "caniuse-lite": "^1.0.30001660",
     "chart.js": "3.4.1",
     "chroma-js": "^2.4.2",
     "country-state-city": "^3.1.2",
@@ -38,11 +39,11 @@
     "npm-watch": "^0.11.0",
     "prop-types": "^15.8.1",
     "pusher-js": "^8.4.0-rc2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-chartjs-2": "3.0.4",
     "react-circular-slider-svg": "0.1.5",
     "react-data-table-component": "^7.5.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-dropzone": "^14.2.2",
     "react-dropzone-component": "^3.2.0",
     "react-github-btn": "^1.4.0",
