@@ -1,7 +1,5 @@
-import Layout from "@/components/layout/Layout";
 import Banner from "@/components/sections/Banner";
 import BannerHome1 from "@/components/sections/BannerHome1";
-import News2 from "@/components/sections/News2";
 import Payments2 from "@/components/sections/Payments2";
 import Testimonials2 from "@/components/sections/Testimonials2";
 import WhyTravelUs from "@/components/sections/WhyTravelUs";
@@ -9,8 +7,8 @@ import ExclusiveService from "@/components/sections/ExclusiveService";
 import TopRatedHotels from "@/components/sections/TopRatedHotels";
 import { ExperienceInverano } from "@/components/sections/ExperienceInverano";
 import { ExperienceVerano } from "@/components/sections/ExperienceVerano";
-import Header from "@/components/layout/header/Header";
 import MasterLayout from "@/components/layout/MasterLayout";
+import HotelesTopRated from "@/components/sections/HotelsTopRated";
 
 export default function Home() {
   return (
@@ -22,7 +20,7 @@ export default function Home() {
         {/* <ExperienceInvirenoSlider /> */}
         <Banner />
         <ExclusiveService />
-        <TopRatedHotels />
+        <HotelesTopRated />
         <Payments2 />
         <WhyTravelUs />
         {/* <RecentLauchedCar /> */}
